move executionstate off of smart contract
