+++ src/lib/global-context.ts
@@ -1,15 +1,63 @@
-import { Party } from './party';
+import * as Mina from './mina';
+import { Body, Party } from './party';
+import { SmartContract } from './zkapp';
 
 export {
+  getExecutionState,
   withContext,
   withContextAsync,
   getContext,
-  mainContext,
   inProver,
   inCompile,
   inCheckedComputation,
 };
 
+// per-smart-contract context for transaction construction
+type ExecutionState = {
+  transactionId: number;
+  partyIndex: number;
+  party: Party;
+};
+
+function getExecutionState(smartContract: SmartContract): ExecutionState {
+  // TODO reconcile mainContext with currentTransaction
+  if (mainContext !== undefined) {
+    return {
+      transactionId: 0,
+      partyIndex: 0,
+      party: mainContext.self,
+    };
+  }
+
+  if (Mina.currentTransaction === undefined) {
+    throw new Error('Cannot execute outside of a Mina.transaction() block.');
+  }
+
+  let executionState = executionStates.get(smartContract);
+  if (
+    executionState !== undefined &&
+    executionState.transactionId === Mina.nextTransactionId.value
+  ) {
+    return executionState;
+  }
+  let id = Mina.nextTransactionId.value;
+  let index = Mina.currentTransaction.nextPartyIndex++;
+  let body = Body.keepAll(smartContract.address);
+  let party = new Party(body);
+  Mina.currentTransaction.parties.push(party);
+
+  executionState = {
+    transactionId: id,
+    partyIndex: index,
+    party,
+  };
+  executionStates.set(smartContract, executionState);
+  return executionState;
+}
+
+const executionStates = new WeakMap<SmartContract, ExecutionState>();
+
+// context for compiling / proving
 // TODO reconcile mainContext with currentTransaction
 let mainContext = undefined as
   | {
